+++ integrationtests/scenarios/i001538-symbols-src-folder-structure/before/Linked/LinkedInNestedFolder.cs
@@ -1,0 +1,6 @@
+﻿namespace PackWithSource.Folder.NestedFolder
+{
+    class LinkedInNestedFolder
+    {
+    }
+}